chore(rust): Set default toolchain
Inspired by this blogpost (https://swatinem.de/blog/rust-toolchain/)
we are setting our rust toolchain and will
start managing it as a dependency. This will make compiling more easy in
case this project is ever shelved. Additionally, once used in CI it will
ensure that CI should not randomly break when new clippy checks land in
the latest rust version for example.