feat(xtask): Add xtasks
xtasks allow us to write rust implementations for common tasks that
happen around the development of Brane. The advantages of writing these
tasks in rust itself, is that is does not pose any additional
dependencies besides the once that we already have.

Additionally, it provides us with the full power of the rust language we
in automating these common tasks

Inspired by: https://github.com/matklad/cargo-xtask

We start off with two subcommands: completions and manpage generation

This also allows us to remove the pesky completion binaries from the
brane-ctl and brane-cli workspace members.